refactor: did miscellaneous refactoring
